some quick cleanups, consolidated in this commit
- make __tests__ folders for tests
- remove the call to prettier on adding a widget (we shouldn't generate non-prettified code)
- delete store.ts, revisit later
- add "shared" to tsconfig.json
