+++ packages/cra-template-modular-typescript/template/src/__tests__/App.test.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { render } from '@testing-library/react';
-import App from './App';
+import App from '../App';
 
 test('renders learn react link', () => {
   const { getByText } = render(<App />);
