+++ packages/create-modular-react-app/template/tsconfig.json
@@ -15,5 +15,5 @@
     "noEmit": true,
     "jsx": "react"
   },
-  "include": ["app", "widgets"]
+  "include": ["app", "widgets", "shared"]
 }
