+++ packages/modular-template-widget-typescript/template/__tests__/index.test.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { render } from 'react-dom';
-import ComponentName__ from './index';
+import ComponentName__ from '../index';
 
 test('it should render', () => {
   const el = document.createElement('div');
